move other structs and impls in their desired location
